временно

+++ TravelAgency/TravelAgencyBusinessLogic/BindingModels/NumberofhotelBindingModel.cs
@@ -10,8 +10,8 @@ namespace TravelAgencyBusinessLogic.BindingModels
         public string Viewnumber { get; set; }
         public DateTime Datearrival { get; set; }
         public DateTime Dateofdeparture { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public int Typeofnumberid { get; set; }
-        public Dictionary<int, int> HotelNumberofhotel { get; set; }
+        public Dictionary<int, decimal> HotelNumberofhotel { get; set; }
     }
 }
