+++ TravelAgency/TravelAgencyBusinessLogic/ViewModels/NumberofhotelViewModel.cs
@@ -10,9 +10,9 @@ namespace TravelAgencyBusinessLogic.ViewModels
         public string Viewnumber { get; set; }
         public DateTime Datearrival { get; set; }
         public DateTime Dateofdeparture { get; set; }
-        public int Price { get; set; }
+        public decimal Price { get; set; }
         public string Type { get; set; }
         public int Typeofnumberid { get; set; }
-        public Dictionary<int, int> HotelNumberofhotel { get; set; }
+        public Dictionary<int, decimal> HotelNumberofhotel { get; set; }
     }
 }
